Deprecate Images.decodeImageFromPixels()
