+++ packages/flame/lib/src/extensions/image.dart
@@ -1,14 +1,34 @@
+import 'dart:async';
 import 'dart:typed_data';
 import 'dart:ui';
 
-import '../flame.dart';
 import 'color.dart';
 import 'vector2.dart';
 
 export 'dart:ui' show Image;
 
 extension ImageExtension on Image {
-  /// Helper method for retrieve the pixel data in a Uint8 format.
+  /// Converts a raw list of pixel values into an [Image] object.
+  ///
+  /// The pixels must be in the RGBA format, i.e. first 4 bytes encode the red,
+  /// green, blue, and alpha components of the first pixel, next 4 bytes encode
+  /// the next pixel, and so on. The pixels are in the row-major order, meaning
+  /// that first [width] pixels encode the first row of the image, next [width]
+  /// pixels the second row, and so on.
+  static Future<Image> fromPixels(Uint8List pixels, int width, int height) {
+    assert(pixels.length == width * height * 4);
+    final completer = Completer<Image>();
+    decodeImageFromPixels(
+      pixels,
+      width,
+      height,
+      PixelFormat.rgba8888,
+      completer.complete,
+    );
+    return completer.future;
+  }
+
+  /// Helper method to retrieve the pixel data of the image as a [Uint8List].
   ///
   /// Pixel order used the [ImageByteFormat.rawRgba] meaning it is: R G B A.
   Future<Uint8List> pixelsInUint8() async {
@@ -43,8 +63,7 @@ extension ImageExtension on Image {
       newPixelData[i + 2] = color.blue;
       newPixelData[i + 3] = color.alpha;
     }
-
-    return Flame.images.decodeImageFromPixels(newPixelData, width, height);
+    return fromPixels(newPixelData, width, height);
   }
 
   /// Change each pixel's color to be brighter and return a new [Image].
@@ -69,7 +88,6 @@ extension ImageExtension on Image {
       newPixelData[i + 2] = color.blue;
       newPixelData[i + 3] = color.alpha;
     }
-
-    return Flame.images.decodeImageFromPixels(newPixelData, width, height);
+    return fromPixels(newPixelData, width, height);
   }
 }
