+++ packages/flame/test/extensions/image_extension_test.dart
@@ -1,0 +1,24 @@
+import 'dart:typed_data';
+
+import 'package:flame/extensions.dart';
+import 'package:test/test.dart';
+
+final output = List.filled(8 * 8 * 4, 255);
+
+void main() {
+  group('ImageExtension', () {
+    test('fromPixels', () async {
+      final data = Uint8List(8 * 8 * 4);
+      for (var i = 0; i < data.length; i += 4) {
+        data[i] = 255;
+        data[i + 1] = 255;
+        data[i + 2] = 255;
+        data[i + 3] = 255;
+      }
+      final image = await ImageExtension.fromPixels(data, 8, 8);
+      final bytes = await image.toByteData();
+
+      expect(bytes!.buffer.asUint8List(), equals(output));
+    });
+  });
+}
